set delete all button
