+++ src/index.html
@@ -21,7 +21,7 @@
            
         </ul>
         <div class="remove">
-            <button type="reset">Clear all completed</button>
+            <button type="submit" class="removeCompleted">Clear all completed</button>
         </div>
  
     </div>
