+++ src/index.js
@@ -1,61 +1,7 @@
 // import _ from 'lodash';
 import './style.css';
-
-class ToDolist {
-#toDo;
-
-constructor() {
-  this.#toDo = [];
-}
-
-// Setters
-SetToDo(toDo) {
-  this.#toDo = toDo;
-}
-
-// Getter
-get ToDo() {
-  return this.#toDo;
-}
-
-reorder() {
-  for (let index = 0; index < this.#toDo.length; index += 1) {
-    this.#toDo[index].id = index;
-  }
-}
-
-// editToDo(id, activity, doneBox) {
-// }
-
-// Methods
-AddToDo(activity, doneBox) {
-  const id = this.#toDo.length;
-  this.#toDo.push({
-    activity,
-    doneBox,
-    id,
-  });
-}
-
-DeleteToDo(id) {
-  // const localToDo = this.#toDo;
-  this.#toDo.splice(id, 1);
-}
-
-SaveToDolistLocal() {
-  localStorage.setItem('toDo', JSON.stringify(this.#toDo));
-}
-
-LoadToDoFromLocal() {
-  const savedToDo = JSON.parse(localStorage.getItem('toDo'));
-
-  if (Array.isArray(savedToDo)) {
-    this.#toDo = savedToDo;
-    return true;
-  }
-  return false;
-}
-}
+import ToDolist from './toDos';
+const removeCompleted = document.querySelector('.removeCompleted');
 
 const myToDolist = new ToDolist();
 
@@ -94,9 +40,7 @@ const render = () => {
     }
 
     inputTask.addEventListener('change', () => {
-      // if (event.key === 'Enter'){
       modifyDescription(index, inputTask.value);
-      // }
     });
 
     doneBox.addEventListener('click', () => {
@@ -128,6 +72,8 @@ const render = () => {
     deleteButton.id = toDo.id;
     element.appendChild(deleteButton);
   }
+
+  
 };
 
 render();
@@ -140,4 +86,13 @@ button.addEventListener('click', () => {
   myToDolist.SaveToDolistLocal();
   titletextbox.value = '';
   render();
+});
+
+removeCompleted.addEventListener('click', () => {
+const temp = myToDolist.ToDo.filter(({doneBox})=> !doneBox);
+
+myToDolist.SetToDo(temp);
+myToDolist.reorder();
+myToDolist.SaveToDolistLocal();
+render(); 
 });
